+++ models/mobilenetv2/mobilenetv2_ssd_fpnlite_640_config.json
@@ -1,0 +1,34 @@
+{
+  "model": {
+    "name": "MobileNetV2-SSD-FPNLite-640",
+    "image_size": 640,
+    "num_channels": 3,
+    "num_classes": 91,
+    "num_anchors": 6,
+    "n_levels": 5,
+    "n_tower": 4
+  },
+  "backbone": {
+    "stem_out_channels": 32,
+    "head_out_channels": 1280,
+    "c3_block_idx": 5,
+    "c4_block_idx": 12,
+    "inverted_residual_setting": [
+      [1,  16, 1, 1],
+      [6,  24, 2, 2],
+      [6,  32, 3, 2],
+      [6,  64, 4, 2],
+      [6,  96, 3, 1],
+      [6, 160, 3, 2],
+      [6, 320, 1, 1]
+    ]
+  },
+  "fpn": {
+    "channels": 128
+  },
+  "paths": {
+    "tf_ckpt_name": "ssd_mobilenet_v2_fpnlite_640x640_coco17_tpu-8",
+    "tf_ckpt_url": "http://download.tensorflow.org/models/object_detection/tf2/20200711/ssd_mobilenet_v2_fpnlite_640x640_coco17_tpu-8.tar.gz",
+    "bin_dir": "mobilenetv2_ssd_fpnlite_bin"
+  }
+}
